requirements.txt
click==8.1.3
Pillow==9.4.0